refactor: NavigationBar 컴포넌트 분리
