+++ app/exhibition/[exhibitionId]/components/NavigationBar/NavigationBar.styles.tsx
@@ -1,0 +1,9 @@
+import NavigationBar from "@/components/NavigationBar/NavigationBar";
+import styled from "styled-components";
+
+export const NavigationBarStyled = styled(NavigationBar)`
+  position: absolute;
+  width: 100%;
+  padding-top: 50px;
+  z-index: 100;
+`;
